chatbot/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.0.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
faster-whisper==1.2.0
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
onnxruntime==1.22.1
openai==1.102.0
packaging==25.0
pillow==11.0.0
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sniffio==1.3.1
soundfile==0.13.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
